Make FrameTime part of the engine interface
