+++ example_plugins/shader/src/lib.rs
@@ -4,9 +4,9 @@ use cimvr_common::{
         Mesh, MeshHandle, Primitive, Render, RenderExtra, ShaderHandle, ShaderSource, UploadMesh,
         Vertex,
     },
-    FrameTime, Transform,
+    Transform,
 };
-use cimvr_engine_interface::{dbg, make_app_state, pkg_namespace, prelude::*};
+use cimvr_engine_interface::{dbg, make_app_state, pkg_namespace, prelude::*, FrameTime};
 
 struct ServerState {
     cube_ent: EntityId,
